+++ tests/functional/test__big_p.py
@@ -28,11 +28,14 @@ class Test_P(unittest.ResetRegisters, unittest.FunctionalTestCase):
         self.register('"abc')
         self.eq('x|y', 'P', 'xab|cy')
         self.assertRegister('"abc')
+        self.eq('x|y', '2P', 'xabcab|cy')
+        self.eq('x|y', '3P', 'xabcabcab|cy')
 
     def test_P_paste_characterwise_content_multiline(self):
         self.register('"fizz\nbuzz')
         self.eq('x|y', 'P', 'x|fizz\nbuzzy')
         self.assertRegister('"fizz\nbuzz')
+        self.eq('x|y', '3P', 'x|fizz\nbuzzfizz\nbuzzfizz\nbuzzy')
 
     def test_P_paste_characterwise_content_with_newline(self):
         self.register('"abc\n')
@@ -48,6 +51,7 @@ class Test_P(unittest.ResetRegisters, unittest.FunctionalTestCase):
         self.register('"abc\n', linewise=True)
         self.eq('x\ny|z', 'P', 'x\n|abc\nyz')
         self.assertRegister('"abc\n', linewise=True)
+        self.eq('a\n|b\nc', '3P', 'a\n|abc\nabc\nabc\nb\nc')
 
     def test_P_paste_linewise_content_multiline(self):
         self.register('"fizz\nbuzz\n', linewise=True)
@@ -88,6 +92,32 @@ class Test_P(unittest.ResetRegisters, unittest.FunctionalTestCase):
         self.register('"bc\n')
         self.eq('abc\nd|ef', 'P', 'abc\nd|bc\nef')
         self.assertRegister('"bc\n')
+
+    def test_paste_multiple_cursor_content(self):
+        self.register('"', ['zz bu'])
+        self.eq('fi|zz', 'P', 'fizz b|uzz')
+        self.register('"', ['zz', 'bu'])
+        self.eq('fi| x |zz', 'P', 'fiz|z x b|uzz')
+        self.register('"', ['on', 'tw', 'thre'])
+        self.eq('x 1|e 2|o 3|e x', 'P', 'x 1o|ne 2t|wo 3thr|ee x')
+
+    def test_v_paste_multiple_cursor_content(self):
+        self.register('"', ['zz bu'])
+        self.eq('x|iz|y', 'v_P', 'n_xzz b|uy')
+        self.register('"', ['one', 'two'])
+        self.eq('a|xx|b a|xx|b', 'v_P', 'n_aon|eb atw|ob')
+        self.register('"', ['on', 'tw', 'thre'])
+        self.eq('x 1|xyz|e 2|xyz|o 3|xyz|e x', 'v_P', 'n_x 1o|ne 2t|wo 3thr|ee x')
+
+    @unittest.mock_bell()
+    def test_paste_multiple_cursor_content_count_not_equal_rings_bell(self):
+        self.register('"', ['on', 'tw'])
+        self.eq('x 1|e 2|o 3|e x', 'P', 'x 1|e 2|o 3|e x')
+        self.assertBell()
+
+    def test_paste_multiple_cursor_content_into_one(self):
+        self.register('"', ['izz', 'buz'])
+        self.eq('f|z', 'P', 'fizzbu|zz')
 
     @unittest.mock_status_message()
     def test_nothing_in_register(self):